pyproject.toml
[project]
name = "magicnacki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "pygame-ce>=2.5.5",
    "requests>=2.32.5",
]
